+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.9.0] - 2023-05-03
+
+### Added
+
+* Counter `EntryPoints_Tas_Parses`, to count per entrypoint, how many parses have been done.
+  It allows to identify, which service components are generating most distinct queries and thus bloating the cache or just burning CPU for parses.
+
+### Changed
+
+* DAS for database access done outside of entrypoints are now collected in a separate thread.
+  This would allow to collect those metrics even when the service does not have any entrypoints or they are
+  infrequently accessed.
+
 ## [2.8.3] - 2023-03-16
 
 ### Fixed